02-Cart Analyzer/requirements.txt
selenium==4.34.2
load-dotenv==0.1.0
streamlit==1.47.0
openai==1.97.1